Refactored AdditionalResources into its own component. Updated link for CareerScope.
